+++ osu.Game.Rulesets.Sentakki/SimaiSharp/Structures/MaiChart.cs
@@ -6,7 +6,7 @@ namespace SimaiSharp.Structures
     [Serializable]
     public sealed class MaiChart
     {
-        public float? FinishTiming { get; set; }
+        public double? FinishTiming { get; set; }
         public NoteCollection[] NoteCollections { get; set; } = Array.Empty<NoteCollection>();
         public TimingChange[] TimingChanges { get; set; } = Array.Empty<TimingChange>();
     }
